First pass at getting tempest integrated with rcbops chef-cookbooks
This pull request adds an image upload service to the cookbook
to add an image that is owned by the tempest monitoring user and will
be used for tempest tests.  This also corrects the tempest git checkout
to use the correct branches for the varying versions of deployed
OpenStack.

!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
This is still in development and a lot of this code base
needs to be simplified to make this generic and portable.
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!

The testing users are created in keystone correctly
on first run so tere is no additional information that needs to be
entered to make this work.

For development I am placing this on a node that is running
running keystone and glance.  There are still some dependency
items that need to be ironed out so you can run this on a node that
does not contain those services in their run_list.

BUGS:
At the moment you have to run chef-client on the node twice so we
can populate the image_id for the test image correctly.  If you
do not run chef-client twice on the node then you will end up
with an image_reg in the tempest.conf configuration file that is
empty.